pyproject.toml
[project]
name = "forex"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "dbt-bigquery>=1.10.2",
    "dlt[bigquery]>=0.4.0",
    "google-cloud-bigquery>=3.26.0",
    "google-cloud-bigquery-storage>=2.27.0",
    "load-dotenv>=0.1.0",
    "requests>=2.32.5",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.6.9",
    "ty>=0.0.1",
]
